requirements.txt
Flask==3.0.3
Werkzeug==3.0.3
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.2
XlsxWriter==3.2.0
python-dotenv==1.0.1
requests==2.31.0
gunicorn==22.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.8.0
argon2-cffi==23.1.0
